requirements.txt
gradio
google-generativeai
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pypdf
python-dotenv